requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
future==0.18.2
idna==2.10
iso8601==0.1.14
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
paho-mqtt==1.5.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
setuptools==57.0.0
urllib3==1.26.5
Werkzeug==2.0.1
